refactor(...) : dto 객체의 패키지 구조 변경
