+++ src/main/java/lotto/controller/LottoController.java
@@ -3,10 +3,10 @@ package lotto.controller;
 import java.util.Map.Entry;
 import lotto.domain.Prize;
 import lotto.domain.PrizeResult;
-import lotto.domain.RandomLottos;
-import lotto.domain.WinningLotto;
 import lotto.domain.model.Lotto;
 import lotto.domain.model.LottoNumber;
+import lotto.dto.RandomLottos;
+import lotto.dto.WinningLotto;
 import lotto.exception.model.LottoExceptionBase;
 import lotto.service.LottoService;
 import lotto.util.Parser;
