+++ src/main/java/lotto/dto/WinningLotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.dto;
 
 import java.util.List;
 import lotto.domain.model.Lotto;
