+++ src/main/java/lotto/service/LottoService.java
@@ -8,10 +8,10 @@ import java.util.Set;
 import java.util.stream.Stream;
 import lotto.domain.Prize;
 import lotto.domain.PrizeResult;
-import lotto.domain.RandomLottos;
-import lotto.domain.WinningLotto;
 import lotto.domain.model.Lotto;
 import lotto.domain.model.LottoNumber;
+import lotto.dto.RandomLottos;
+import lotto.dto.WinningLotto;
 import lotto.exception.LottoNumberException;
 import lotto.exception.message.Error;
 import lotto.util.LottoMaker;
